chore: rustfmt, move tests to neon/mod.rs
